projects/Windows: remove wolfSSL from legacy projects
- Remove wolfSSL from the legacy projects for Visual Studio 2010 - 2013.

It's no longer possible to maintain the custom build configuration of
wolfSSL for these old versions of Visual Studio.

Note support for wolfSSL was recently added to the winbuild build system
in 4de627ab and the user could possibly make their own wolfSSL build and
attempt to link it using an old version of Visual Studio that way.

Ref: https://gist.github.com/jay/5f6d8d5ba15c12c7457e3216a94da72d

Closes https://github.com/curl/curl/pull/15468