GHA/windows: drop running tests with dl-mingw 7.3.0 due to flakiness
This job was never stable. Bumping to -j8 (from -j4) possibly made it
flakier: 032447e6249bf87958b16eb0c97874490b711ec7 #16271

Keep this job for build tests and drop running tests to improve the CI
experience and save CI time.

It's also a simple build with no dependencies. CI continues to build
a similar job with 9.5.0, which is more stable.

It remains a puzzle why builds with this toolchain (7.3.0 win32 threads
mingw-builds) is flakier and requires more test exceptions than the
indentical build with a slightly different build/version of
the toolchain (9.5.0 posix threads winlibs_mingw).

Ref: #14854
Closes #16564